refactor(chalice): prevent SSRF
