+++ api/chalicelib/core/webhook.py
@@ -1,14 +1,8 @@
-import ipaddress
 import logging
-import socket
 from typing import Optional
-from urllib.parse import urlparse, urlunparse
 
-import requests
-from decouple import config
-from requests.adapters import HTTPAdapter
 import schemas
-from chalicelib.utils import pg_client, helper
+from chalicelib.utils import pg_client, helper, ssrf
 from chalicelib.utils.TimeUTC import TimeUTC
 from chalicelib.utils.log import sanitize
 from fastapi import HTTPException, status
@@ -174,71 +168,13 @@ def trigger_batch(data_list):
                 logger.exception(f"!!Error while triggering webhook_id={w['destination']}")
 
 
-def __get_allowed_hosts():
-    return {h.strip().lower() for h in config("WEBHOOK_ALLOWED_HOSTS", default="").split(",") if h.strip()}
-
-
-def __resolve_public_ip(endpoint):
-    """Validates that the endpoint resolves to public IPs only and returns one of them,
-    so the request can be pinned to it (protects against DNS-rebinding between check and request).
-    Returns None (no validation, no pinning) for hosts listed in WEBHOOK_ALLOWED_HOSTS."""
-    parsed = urlparse(endpoint)
-    hostname = parsed.hostname
-    if not hostname:
-        raise ValueError(f"webhook endpoint has no valid hostname: {endpoint}")
-    if hostname.lower() in __get_allowed_hosts():
-        return None
-    try:
-        resolved = socket.getaddrinfo(hostname, parsed.port or (443 if parsed.scheme == "https" else 80),
-                                      proto=socket.IPPROTO_TCP)
-    except socket.gaierror as e:
-        raise ValueError(f"webhook endpoint hostname could not be resolved: {hostname}") from e
-    ips = []
-    for family, _, _, _, sockaddr in resolved:
-        ip = ipaddress.ip_address(sockaddr[0])
-        if not ip.is_global:
-            raise ValueError(f"webhook endpoint resolves to a non-public IP address: {hostname} -> {ip}")
-        ips.append(ip)
-    return str(ips[0])
-
-
-class _PinnedHostHTTPSAdapter(HTTPAdapter):
-    """Keeps TLS SNI and certificate validation bound to the original hostname
-    while the connection itself targets a pre-resolved IP."""
-
-    def __init__(self, hostname, **kwargs):
-        self._hostname = hostname
-        super().__init__(**kwargs)
-
-    def init_poolmanager(self, *args, **kwargs):
-        kwargs["server_hostname"] = self._hostname
-        kwargs["assert_hostname"] = self._hostname
-        super().init_poolmanager(*args, **kwargs)
-
-
-def __post(endpoint, pinned_ip, json_data, headers):
-    if pinned_ip is None:
-        return requests.post(url=endpoint, json=json_data, headers=headers, allow_redirects=False)
-    parsed = urlparse(endpoint)
-    hostname = parsed.hostname
-    ip_host = f"[{pinned_ip}]" if ":" in pinned_ip else pinned_ip
-    netloc = ip_host if parsed.port is None else f"{ip_host}:{parsed.port}"
-    pinned_url = urlunparse(parsed._replace(netloc=netloc))
-    headers = {**headers, "Host": hostname if parsed.port is None else f"{hostname}:{parsed.port}"}
-    with requests.Session() as s:
-        if parsed.scheme == "https":
-            s.mount("https://", _PinnedHostHTTPSAdapter(hostname))
-        return s.post(url=pinned_url, json=json_data, headers=headers, allow_redirects=False)
-
-
 def __trigger(hook, data):
     if hook is not None and hook["type"] == 'webhook':
-        pinned_ip = __resolve_public_ip(hook["endpoint"])
         headers = {}
         if hook["authHeader"] is not None and len(hook["authHeader"]) > 0:
             headers = {"Authorization": hook["authHeader"]}
 
-        r = __post(endpoint=hook["endpoint"], pinned_ip=pinned_ip, json_data=data, headers=headers)
+        r = ssrf.post_json(endpoint=hook["endpoint"], json_data=data, headers=headers)
         if r.status_code != 200:
             logger.error("=======> webhook: something went wrong for:")
             logger.error(hook)
